pyproject.toml
[tool.poetry]
name = "pan-scm-sdk"
version = "0.1.0"
description = "Python SDK for Palo Alto Networks Strata Cloud Manager."
authors = ["Calvin Remsburg <calvin@cdot.io>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jwt = "^1.3.1"
oauthlib = "^3.2.2"
requests-oauthlib = "^2.0.0"
setuptools = "^75.1.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
mkdocs = "^1.6.1"
mkdocs-material = "^9.5.39"
mkdocstrings = "^0.26.1"
pytest = "^8.3.3"
factory-boy = "^3.3.1"
termynal = "^0.12.1"
invoke = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
